Feat : Phone verification feature test on Patient Model

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Middleware\ConfirmPhonePatient;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -23,14 +24,23 @@ Route::get('/about','User\AboutController@index')->name('user.about');
 
 // Routes for client patient
 Route::prefix('/patient')->namespace('Patient')->name('patient.')->group(function() {
-    Route::get('/home', 'PatientController@index')->name('home');
-    Route::get('/profile/{id}', 'PatientController@profile')->name('profile');
-    Route::put('/profile/{id}', 'PatientController@update')->name('update');
-    Route::patch('/profile/{id}', 'PatientController@email')->name('email');
-    Route::put('/profil/{id}', 'PatientController@password')->name('password');
-    Route::delete('/profile/{id}', 'PatientController@destroy')->name('destroy');
 
-    Route::get('/identifiant/{id}','IdentifiantController@index')->name('identifiant');
+    // patient confirmation phone number
+    Route::get('/confirm-phone', 'PatientController@confirmPhonePage')->name('confirm.tampon');
+    Route::post('/confirm-phone', 'PatientController@confirmPhone')->name('confirm');
+
+    Route::middleware([ConfirmPhonePatient::class])->group(function() {
+        Route::get('/home', 'PatientController@index')->name('home');
+        Route::get('/profile/{id}', 'PatientController@profile')->name('profile');
+        Route::put('/profile/{id}', 'PatientController@update')->name('update');
+        Route::patch('/profile/{id}', 'PatientController@email')->name('email');
+        Route::put('/profil/{id}', 'PatientController@password')->name('password');
+        Route::delete('/profile/{id}', 'PatientController@destroy')->name('destroy');
+    
+        Route::get('/identifiant/{id}','IdentifiantController@index')->name('identifiant');
+    });
+
+    // Authentification des
     Route::namespace('Auth')->group(function(){
         Route::get('/login', 'LoginController@showLoginForm')->name('login');
         Route::post('/login', 'LoginController@login')->name('login');
